+++ bugsnag/src/main/java/com/bugsnag/delivery/Delivery.java
@@ -1,6 +1,6 @@
 package com.bugsnag.delivery;
 
-import com.bugsnag.serialization.ISerializer;
+import com.bugsnag.serialization.Serializer;
 
 import java.util.Map;
 
@@ -12,7 +12,7 @@ public interface Delivery {
      * @param object     the object to deliver.
      * @param headers    HTTP headers which must be appended to the request.
      */
-    void deliver(ISerializer serializer, Object object, Map<String, String> headers);
+    void deliver(Serializer serializer, Object object, Map<String, String> headers);
 
     /**
      * Close any open connections to Bugsnag.
