+++ bugsnag/src/main/java/com/bugsnag/delivery/OutputStreamDelivery.java
@@ -1,6 +1,6 @@
 package com.bugsnag.delivery;
 
-import com.bugsnag.serialization.ISerializer;
+import com.bugsnag.serialization.Serializer;
 import com.bugsnag.serialization.SerializationException;
 
 import org.slf4j.Logger;
@@ -18,7 +18,7 @@ public class OutputStreamDelivery implements Delivery {
     }
 
     @Override
-    public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+    public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
         try {
             serializer.writeToStream(outputStream, object);
         } catch (SerializationException ex) {
