+++ bugsnag/src/test/java/com/bugsnag/BugsnagTest.java
@@ -10,7 +10,7 @@ import com.bugsnag.callbacks.Callback;
 import com.bugsnag.delivery.Delivery;
 import com.bugsnag.delivery.HttpDelivery;
 import com.bugsnag.delivery.OutputStreamDelivery;
-import com.bugsnag.serialization.ISerializer;
+import com.bugsnag.serialization.Serializer;
 
 import org.junit.After;
 import org.junit.Before;
@@ -99,7 +99,7 @@ public class BugsnagTest {
     public void testProjectPackages() {
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertTrue(report.getExceptions().get(0).getStacktrace().get(0).isInProject());
             }
@@ -118,7 +118,7 @@ public class BugsnagTest {
         bugsnag.setAppVersion("1.2.3");
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("1.2.3", report.getApp().get("version"));
             }
@@ -135,7 +135,7 @@ public class BugsnagTest {
         bugsnag.setAppType("testtype");
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("testtype", report.getApp().get("type"));
             }
@@ -151,7 +151,7 @@ public class BugsnagTest {
     public void testSeverity() {
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals(Severity.INFO.getValue(), report.getSeverity());
             }
@@ -169,7 +169,7 @@ public class BugsnagTest {
         bugsnag.setDelivery(new Delivery() {
             @SuppressWarnings("unchecked")
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 Map<String, Object> firstTab =
                         (Map<String, Object>) report.getMetaData().get("firsttab");
@@ -201,7 +201,7 @@ public class BugsnagTest {
         bugsnag.setDelivery(new Delivery() {
             @SuppressWarnings("unchecked")
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 Map<String, Object> requestTab =
                         (Map<String, Object>) report.getMetaData().get("request");
@@ -238,7 +238,7 @@ public class BugsnagTest {
     public void testUser() {
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("123", report.getUser().get("id"));
                 assertEquals("test@example.com", report.getUser().get("email"));
@@ -267,7 +267,7 @@ public class BugsnagTest {
         });
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("the context", report.getContext());
             }
@@ -289,7 +289,7 @@ public class BugsnagTest {
         });
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("the grouping hash", report.getGroupingHash());
             }
@@ -311,7 +311,7 @@ public class BugsnagTest {
         });
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("newapikey", report.getApiKey());
             }
@@ -327,7 +327,7 @@ public class BugsnagTest {
     public void testSingleCallbackInNotify() {
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("newapikey", report.getApiKey());
             }
@@ -361,7 +361,7 @@ public class BugsnagTest {
         });
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertEquals("secondnewapikey", report.getApiKey());
             }
@@ -406,7 +406,7 @@ public class BugsnagTest {
             }
 
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 assertEquals("https://www.example.com", endpoint);
             }
 
@@ -438,7 +438,7 @@ public class BugsnagTest {
             }
 
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 assertEquals("/127.0.0.1:8080", proxy.address().toString());
             }
 
@@ -457,7 +457,7 @@ public class BugsnagTest {
         bugsnag.setSendThreads(true);
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 // There is information about at least one thread
                 assertTrue(report.getThreads().size() > 0);
@@ -474,7 +474,7 @@ public class BugsnagTest {
     public void testHandledIncrementNoSession() {
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
                 assertNull(report.getSession());
             }
@@ -491,7 +491,7 @@ public class BugsnagTest {
         bugsnag.startSession();
         bugsnag.setDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
                 Report report = ((Notification) object).getEvents().get(0);
 
                 Map<String, Object> session = report.getSession();
