+++ features/fixtures/scenarios/src/main/java/com/bugsnag/mazerunner/scenarios/HandledSessionScenario.java
@@ -2,7 +2,7 @@ package com.bugsnag.mazerunner.scenarios;
 
 import com.bugsnag.Bugsnag;
 import com.bugsnag.delivery.Delivery;
-import com.bugsnag.serialization.ISerializer;
+import com.bugsnag.serialization.Serializer;
 
 import java.util.Map;
 
@@ -22,7 +22,7 @@ public class HandledSessionScenario extends Scenario {
         // Stop sessions being sent to Bugsnag for this case
         bugsnag.setSessionDelivery(new Delivery() {
             @Override
-            public void deliver(ISerializer serializer, Object object, Map<String, String> headers) {
+            public void deliver(Serializer serializer, Object object, Map<String, String> headers) {
 
             }
